requirements.txt
python-dotenv==1.0.0
tda-api==1.6.0
tweepy==4.14.0
openai==1.3.7
pandas==2.1.3
numpy==1.24.3
python-dateutil==2.8.2
schedule==1.2.1
flask==3.0.0
flask-socketio==5.3.6
flask-login==0.6.3
werkzeug==3.0.1
pyOpenSSL==24.0.0
plotly==5.18.0
gunicorn==21.2.0
eventlet==0.35.2
